# common lisp: update code, speedup