it seems to work faster on my machine (~30-40% faster)