@jinyus 2025-05-29
This is a little faster on the initial input but gets significantly slower as the input size grows.

Main:
```bash
| Common Lisp (SBCL) | 154.00 ms | 1.34 s | 11.20 s | 12.69 s |
```

This PR:
```bash
| Common Lisp (SBCL) | 142.00 ms | 4.03 s | 29.64 s | 33.81 s |
```