@zupat 2026-02-10
Closing this since #553 provides a great improvement